Refactoring AssetWatcher and FileSystemWatcherManager
