--- Editor/Core/Components/AssetWatcher.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5812328d8195a4e1c8b95470930a08ce
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
